footvolley-orchestrator/requirements.txt
mysql-connector-python
python-dotenv
pyyaml
